% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObxContextObject.R
\name{createObxContext}
\alias{createObxContext}
\title{Create an OBX Context.}
\usage{
createObxContext(tdrActiveContext, cpyname)
}
\value{
An `obxActiveContext` object
}
\description{
Create a container (`azureObxContext`) for holding variables used by the `inndxrs` package.
}
