% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdrContextObject.R
\name{createTdrContext}
\alias{createTdrContext}
\title{Create an TDR Context.}
\usage{
createTdrContext(driver, serverurl, database, uid, pwd, port, configFile)
}
\arguments{
\item{driver}{The driver}

\item{serverurl}{The serverurl}

\item{database}{The TDR database}

\item{uid}{Resource uid}

\item{pwd}{The pwd}

\item{port}{The port}

\item{configFile}{location of file that contains configuration in JSON format. By default, this location is at `getOption("inndxrs.config")`}
}
\value{
An `tdrActiveContext` object
}
\description{
Create a container (`azureTdrContext`) for holding variables used by the `inndxrs` package.
}
