% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.tdrActiveContext}
\alias{as.tdrActiveContext}
\alias{is.tdrActiveContext}
\title{tdrActiveContext object.}
\usage{
as.tdrActiveContext(x)

is.tdrActiveContext(x)
}
\arguments{
\item{x}{Object to create, test or print}

\item{...}{Ignored}
}
\description{
Functions for creating and displaying information about tdrActiveContext objects.
}
\seealso{
[createTdrContext()]
}
