% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tdr_companyUI.R, R/mod_tdr_serversUI.R
\name{mod_tdr_companyUI}
\alias{mod_tdr_companyUI}
\alias{mod_tdr_company}
\alias{mod_tdr_servers}
\title{mod_tdr_companyUI and mod_tdr_company}
\usage{
mod_tdr_companyUI(id, label = "")

mod_tdr_company(input, output, session, stringsAsFactors = TRUE)

mod_tdr_servers(input, output, session, stringsAsFactors = TRUE)
}
\arguments{
\item{id}{shiny id}

\item{label}{fileInput label}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{stringsAsFactors}{logical: should character vectors be converted to factors?}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{stringsAsFactors}{logical: should character vectors be converted to factors?}
}
\description{
A shiny Module

mod_tdr_company server function

mod_tdr_company server function
}
\examples{
library(shiny)
library(DT)
if (interactive()){
ui <- fluidPage(
  mod_csv_fileInput("fichier"),
DTOutput("tableau")
)

server <- function(input, output, session) {
  data <- callModule(mod_csv_file,"fichier")
  output$tableau <- renderDT({data()})
}

shinyApp(ui, server)
}

}
