% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tdr_serversUI.R
\name{mod_tdr_serversUI}
\alias{mod_tdr_serversUI}
\title{mod_tdr_serversUI and mod_tdr_servers}
\usage{
mod_tdr_serversUI(id, title, body)
}
\arguments{
\item{id}{shiny id}

\item{label}{fileInput label}
}
\description{
A shiny Module
}
\examples{
library(shiny)
library(DT)
if (interactive()){
ui <- fluidPage(
  mod_tdr_serversUI("fichier"),
DTOutput("tableau")
)

server <- function(input, output, session) {
  data <- callModule(mod_csv_file,"fichier")
  output$tableau <- renderDT({data()})
}

shinyApp(ui, server)
}

}
