% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObxAuthenticate.R
\name{obxAuthenticate}
\alias{obxAuthenticate}
\title{Authenticates against TDR Database.}
\usage{
obxAuthenticate(obxActiveContext, obxname)
}
\arguments{
\item{obxActiveContext}{A container used for caching variables used by `inndxrs`, created by [createObxContext()]}

\item{verbose}{If TRUE, prints verbose messages}
}
\value{
If successful, returns TRUE
}
\description{
Authenticates against TDR Database.
}
