% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObxAuthenticate.R
\name{obxAuthenticateDB}
\alias{obxAuthenticateDB}
\title{Switch based on auth types.}
\usage{
obxAuthenticateDB(obxActiveContext, obxname, verbose = FALSE)
}
\arguments{
\item{obxActiveContext}{A container used for caching variables used by `inndxrs`, created by [createObxContext()]}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
If successful, returns TRUE
}
\description{
Switch based on auth types.
}
\seealso{
Other TDR resource functions: \code{\link{tdrAuthenticateDB}}
}
