% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObxAuthenticate.R
\name{obxGetDBConnection}
\alias{obxGetDBConnection}
\title{Get Azure token using RefreshToken}
\usage{
obxGetDBConnection(dbname, db, verbose = FALSE)
}
\value{
If successful, returns TRUE
}
\description{
Get Azure token using RefreshToken
}
