% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read.inndxrs.config}
\alias{read.inndxrs.config}
\title{Reads settings from configuration file in JSON format.}
\usage{
read.inndxrs.config(configFile = getOption("inndxrs.config"))
}
\arguments{
\item{configFile}{location of file that contains configuration in JSON format. By default, this location is at `getOption("inndxrs.config")`}
}
\description{
Reads settings from configuration file in JSON format.
}
