% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObxContextObject.R
\name{setObxContext}
\alias{setObxContext}
\title{Updates tdrActiveContext object.}
\usage{
setObxContext(obxActiveContext, companyname)
}
\arguments{
\item{obxActiveContext}{A container used for caching variables used by `inndxrs`, created by [createObxContext()]}

\item{companyname}{The port}
}
\description{
Updates the value of an `obxActiveContext` object, created by [createObxContext()]
}
