% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdrContextObject.R
\name{setTdrContext}
\alias{setTdrContext}
\title{Updates tdrActiveContext object.}
\usage{
setTdrContext(tdrActiveContext, driver, serverurl, database, uid, pwd, port)
}
\arguments{
\item{tdrActiveContext}{A container used for caching variables used by `inndxrs`, created by [createTdrContext()]}

\item{driver}{The driver}

\item{serverurl}{The serverurl}

\item{database}{The TDR database}

\item{uid}{Resource uid}

\item{pwd}{The pwd}

\item{port}{The port}
}
\description{
Updates the value of an `tdrActiveContext` object, created by [createTdrContext()]
}
