% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdrAuthenticate.R
\name{tdrAuthenticate}
\alias{tdrAuthenticate}
\title{Authenticates against TDR Database.}
\usage{
tdrAuthenticate(tdrActiveContext, driver, serverurl, database, uid, pwd, port,
  verbose = FALSE)
}
\arguments{
\item{tdrActiveContext}{A container used for caching variables used by `inndxrs`, created by [createTdrContext()]}

\item{driver}{The driver}

\item{serverurl}{The serverurl}

\item{database}{The TDR database}

\item{uid}{Resource uid}

\item{pwd}{The pwd}

\item{port}{The port}

\item{verbose}{If TRUE, prints verbose messages}
}
\value{
If successful, returns TRUE
}
\description{
Authenticates against TDR Database.
}
