% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdrAuthenticate.R
\name{tdrAuthenticateDB}
\alias{tdrAuthenticateDB}
\title{Switch based on auth types.}
\usage{
tdrAuthenticateDB(tdrActiveContext, svrname, verbose = FALSE)
}
\arguments{
\item{tdrActiveContext}{A container used for caching variables used by `inndxrs`, created by [createTdrContext()]}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
If successful, returns TRUE
}
\description{
Switch based on auth types.
}
\seealso{
Other TDR resource functions: \code{\link{obxAuthenticateDB}}
}
