% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdrAuthenticate.R
\name{tdrGetDBConnection}
\alias{tdrGetDBConnection}
\title{Get Azure token using RefreshToken}
\usage{
tdrGetDBConnection(tdrActiveContext, verbose = FALSE)
}
\arguments{
\item{tdrActiveContext}{A container used for caching variables used by `inndxrs`, created by [createTdrContext()]}
}
\value{
If successful, returns TRUE
}
\description{
Get Azure token using RefreshToken
}
